\docType{package}
\name{seedProd-package}
\alias{seedProd-package}
\title{Seed production routines}
\description{
  This package is a collection of different seed production
  routines.
}
\details{
  \tabular{ll}{ Package: \tab seedProd\cr Type: \tab
  Package\cr Version: \tab
  MAJORVERSION.MINORVERSION.PATCHVERSION\cr Date: \tab
  TODAYSDATE\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Rainer M Krug \email{Rainer@krugs.de}
}

